pyproject.toml
[tool.poetry]
name = "thirteen"
version = "0.2.0"
description = "A Discord bot written in Python for my server"
authors = ["kyomi <personal@kyomi.codes>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.1.0"
click = "^8.1.3"
rich = "^12.6.0"
asyncpg = "^0.27.0"
redis = "^4.4.0"
alembic = "^1.9.0"
sqlalchemy = {version = "^1.4.45", extras = ["asyncio"]}

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.12.0"
commitizen = "^2.38.0"

[tool.commitizen]
version = "0.2.0"
version_files = ["pyproject.toml"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
